clippy fixes
